--- GameDataStorage.js
@@ -1,43 +1,0 @@
-export const gameDataStorage = [
-    {
-        id: 1,
-        name: "Battleship",
-        author: "Aaron",
-        thumbnail: "/thumbnail.png",
-        link: "https://battleship-v68e.onrender.com",
-        description: "The classic game made by hasbro that allows a user to play against another person connected to the website.",
-    },
-    {
-        id: 2,
-        name: "Squish-The-Cockroach",
-        author: "Tien Nguyen",
-        thumbnail: "/Squish_the_cockroach.png",
-        link: "https://cockroach-squish.onrender.com/",
-        description: "This game is for anyone who enjoys the calming feel of tap-and-click interactions. In a quick 30-second round, squish as many cockroaches as you can—each squish earns a credit. Log in at the end to save your score. Simple by design, it taps into self-regulation through sensory input: repetitive, small actions that can help ease stress, anxiety, and boredom."
-    },
-    {
-        id: 3,
-        name: "1000 Clicks in 10 minutes",
-        author: "Benedict Antwi",
-        thumbnail: "/1000_clicks.png",
-        link: "https://benedictantwi-gameentry.onrender.com/",
-        description: "The user is challenged to get 1,000 points within a timeframe of 10 minutes. To gain a point, the user can simply click on a button and it will add a point. However, the user can also trade points to gain powerups which could help them get points in a quicker timeframe. While the user will have 10 minutes to complete the challenge, it's best they complete it as fast as possible"
-    },
-    {
-        id: 4,
-        name: "Trust AI?!?!",
-        author: "Aditya Patel",
-        thumbnail: "/thumbnail.png",
-        link: "https://ap-trust-ai.onrender.com",
-        description: "a web-based trivia bluff game, where trivia questions will be generated dynamically using AI so each round is unique. The AI opponent will give answers that may be true or bluff, based on its personality (Honest, Sneaky, or Chaotic). Players will see the question, the AI’s answer, their current score and round, and buttons to Trust or Call Bluff, along with power-ups like Fact Check and Double Down. Points are earned or lost based on correct decisions, and a final score is displayed at the end, creating an interactive challenge of knowledge and intuition."
-    },
-    {
-        id: 5,
-        name: "Unscramble",
-        author: "Aishwarya",
-        thumbnail: "/thumbnail.png",
-        link: "",
-        description: "A word puzzle game where the user has to unscramble a series of letters to identify the real word. The goal of the game is to unscramble as many words as possible within 60 seconds. They can choose between different categories such as food, places, or animals. A score counter and timer will be on display and every word guessed is one point. If a word is guessed incorrectly the user can skip the word or try again. There will be a top 5 leaderboard to track points."
-    }
-    
-]
