+++ index.html
@@ -7,8 +7,8 @@
       <link rel="stylesheet" href="/index.css" />
     <title>Final Project</title>
   </head>
-  <body>
-    <div id="root"></div>
+  <body class="bg-neutral-700">
     <script type="module" src="/src/main.jsx"></script>
+    <div id="root"></div>
   </body>
 </html>
