+++ package.json
@@ -25,6 +25,6 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.4.0",
-    "vite": "^7.1.7"
+    "vite": "^7.1.9"
   }
 }
